Ditch running `lein` to self-install
The idea of this was to make it so that once the box was set up, you did not need internet access, as that is one of the big advantages of Vagrant, allowing you to e.g. code on the train. However, this was running `lein` for the root user not the vagrant user so you still end up self-installing when you get to this point.

I spent a little bit of time trying to get it set up for the vagrant user in the Vagrantfile but couldn't get it working and since I don't need it as I have internet access, I have decided to leave this as a TODO.

If you are reading this and can see how to make it work, please do raise a PR!