Allow remote code execution
For the Actors chapter, we are looking at distributing workloads across multiple computers.

This commit connects the Vagrant box to the public network so that it can connect to another node, and also gives it a hostname for ease of reference.

This commit also adds the Vagrantfile I have put on the other machine to be node 2.