Explain how to access server
This is for when you run the server and the book says it's running at "http://localhost:3000". Normally with Vagrant we would do this with port forwarding, but since the vagrantbox is already on the local network to allow distributing workloads across machines, we don't need to forward ports.

This commit also changes the name of the Vagrant box to be more descriptive and generic.